refactor(hindsight): rename test_support module to test_utils
Matches the repo's test_utils convention. Pure rename of the decoder
test-only module and its imports; no behavior change.
